fix: Add type mappings between Contracts and Domain types
- Fix Attribution handlers: use CreatorCredit instead of CreatorCreditResponse
- Add enum conversions: DifficultyLevel, SessionLength, SubscriptionType, ContributorRole casts
- Add type mappers for CharacterRequest→ScenarioCharacter, SceneRequest→Scene
- Fix AccountSettings mapping from Contracts to Domain
- Fix CharacterMetadata mapping from Dictionary<string,object> to Domain type
- Fix numeric conversions: double→int for PlayerCompassProgressDto.Total, float→int for CurrentScore
- Fix collection conversions: EchoLog/SessionAchievement→object, string[]→List<string>
- Fix SessionStatus enum→string, HealthStatus enum→string
- Fix MusicPalette→string conversion
- Fix MediaAsset→MediaItem mapping
- Use explicit lambda syntax for Select to resolve type inference issues

+++ src/Mystira.App.Application/CQRS/Badges/Queries/GetProfileBadgeProgressQueryHandler.cs
@@ -99,7 +99,7 @@ public sealed class GetProfileBadgeProgressQueryHandler : IQueryHandler<GetProfi
             {
                 CompassAxisId = axisId,
                 CompassAxisName = axisName,
-                CurrentScore = currentScore,
+                CurrentScore = (int)currentScore,
                 Tiers = axisTiers
             });
         }
